+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ## change log ##
 
 * 2016-04-22-25,26: 0.3.1:
+  - v2r: vocab terms is now List[Term]
   - uploaded "owx" (OWL/XML) file is now stored in that format as the original.
   - general adjustments related with OWL/XML: "owx" is the associated request format and also
     the extension of stored file. Mime type is "application/owl+xml". Extension and mime type